Add cut::force, and improved logging format of failed tests
Normally, one would put unit tests along with the code being tested,
so that all relevent unit tests are automatically included by the
linker.  However, sometimes it is desirable to have compilation units
containing nothing but unit tests.  The a static instance of
cut::force will cause the compilation unit to be included, whenever
the library is linked.

Also changed the textual logging of failed tests, to match the emacs
convention for compilation failures, allowing automatic location of
the file/line of each test failure.